Add files via upload
The dataset is a JSON file that contains different tags like greetings, goodbye, hospital_search, pharmacy_search, etc. Each tag contains a list of patterns a user can ask and the responses a chatbot can respond according to that pattern. The dataset is good for understanding how chatbot data works.